+++ docs/source/changelog.rst
@@ -3,6 +3,11 @@ Change Log
 
 .. currentmodule:: idesolver
 
+v1.1.0
+------
+* Changed the name of ``IDESolver.F`` to ``f``.
+* The default global error function is now injected instead of hard-coded.
+* ``solve`` returns `(y, global_error, iteration)` instead of just `y`.
 
 v1.0.0
 ------